+++ amstrax/plugins/events/__init__.py
@@ -1,0 +1,2 @@
+from . import event_processing
+from .event_processing import *
