+++ amstrax/plugins/peaks/__init__.py
@@ -1,0 +1,11 @@
+from . import peak_basics
+from .peak_basics import *
+
+from . import peak_processing
+from .peak_processing import *
+
+from . import radon_peak_basics
+from .radon_peak_basics import *
+
+from . import radon_peaks
+from .radon_peaks import *
