+++ amstrax/plugins/peaks/peak_processing.py
@@ -6,7 +6,7 @@ from immutabledict import immutabledict
 from strax.processing.general import _touching_windows
 
 from amstrax.SiPMdata import *
-from .radon_pulse_processing import HITFINDER_OPTIONS
+from amstrax.plugins.records.pulse_processing import HITFINDER_OPTIONS
 
 export, __all__ = strax.exporter()
 
@@ -61,8 +61,6 @@ export, __all__ = strax.exporter()
     strax.Option('peak_split_min_ratio', default=4,
                  help="Minimum ratio between local sum waveform"
                       "minimum and maxima on either side, to trigger a split"),
-    strax.Option('pmt_channel', default=0,
-                 help="PMT channel for splitting pmt and sipms"), 
     strax.Option('peak_max_duration', default=int(10e6), infer_type=False,
                  help="Maximum duration [ns] of a peaklet"),
     strax.Option('channel_map', track=False, type=immutabledict,
@@ -70,7 +68,7 @@ export, __all__ = strax.exporter()
                       "channel number."),
     *HITFINDER_OPTIONS,
 )
-class RadonPeaks(strax.Plugin):
+class Peaks(strax.Plugin):
     """
     Split records into:
         - peaks
@@ -86,28 +84,28 @@ class RadonPeaks(strax.Plugin):
     the lone_hits includes the left and right hit extension, except the
     extension overlaps with any peak or other hits.
     """
-    __version__ = '0.1.78'
+    __version__ = '0.1.86'
 
     parallel = 'process'
     rechunk_on_save = True
     compressor = 'zstd'
 
-    depends_on = ('radon_records',)
+    depends_on = ('records',)
 
-    provides = ('radon_peaks','radon_lone_hits')
-    data_kind = dict(radon_peaks='radon_peaks',
-                    radon_lone_hits='radon_lone_hits')
+    provides = ('peaks','lone_hits')
+    data_kind = dict(peaks='peaks',
+                    lone_hits='lone_hits')
 
     def infer_dtype(self):
-        return dict(radon_peaks=strax.peak_dtype(
-                        n_channels=4),
-                    radon_lone_hits=strax.hit_dtype)
+        return dict(peaks=strax.peak_dtype(
+                        n_channels=self.config['n_tpc_pmts']),
+                    lone_hits=strax.hit_dtype)
 
-    def compute(self, radon_records, start, end):
+    def compute(self, records, start, end):
         # Hit finding
-        hits = strax.find_hits(radon_records, min_amplitude=amstrax.hit_min_amplitude(self.config['hit_min_amplitude']))
+        hits = strax.find_hits(records, min_amplitude=amstrax.hit_min_amplitude(self.config['hit_min_amplitude']))
 
-        self.to_pe = np.ones(8)
+        self.to_pe = np.ones(self.config['n_tpc_pmts'])
 
         # Remove hits in zero-gain channels
         # they should not affect the clustering!
@@ -123,7 +121,7 @@ class RadonPeaks(strax.Plugin):
             right_extension=self.config['peak_right_extension'],
             min_area=self.config['peak_min_area'],
             min_channels=self.config['peak_min_pmts'],
-            result_dtype=strax.peak_dtype(n_channels=4),
+            result_dtype=strax.peak_dtype(n_channels=self.config['n_tpc_pmts']),
             max_duration=self.config['peak_max_duration'],
         )
 
@@ -138,7 +136,7 @@ class RadonPeaks(strax.Plugin):
         is_lone_hit = strax.fully_contained_in(hits, peaks) == -1
         lone_hits = hits[is_lone_hit]
         strax.integrate_lone_hits(
-            lone_hits, radon_records, peaks,
+            lone_hits, records, peaks,
             save_outside_hits=(self.config['peak_left_extension'],
                                self.config['peak_right_extension']),
             n_channels=len(self.to_pe))
@@ -149,7 +147,7 @@ class RadonPeaks(strax.Plugin):
         # is not extended beyond a peak.
         outside_peaks = self.create_outside_peaks_region(peaks, start, end)
         strax.find_hit_integration_bounds(
-            hits, outside_peaks, radon_records,
+            hits, outside_peaks, records,
             save_outside_hits=(self.config['peak_left_extension'],
                                self.config['peak_right_extension']),
             n_channels=len(self.to_pe),
@@ -167,30 +165,31 @@ class RadonPeaks(strax.Plugin):
         hitlets['time'] = hitlets['time'] - hitlet_time_shift
         hitlets['length'] = (hitlets['right_integration'] - hitlets['left_integration'])
         hitlets = strax.sort_by_time(hitlets)
-        rlinks = strax.record_links(radon_records)
+        rlinks = strax.record_links(records)
 
-        strax.sum_waveform(peaks, hitlets, radon_records, rlinks, self.to_pe)
+        strax.sum_waveform(peaks, hitlets, records, rlinks, self.to_pe)
 
         strax.compute_widths(peaks)
-
         # Split peaks using low-split natural breaks;
         # see https://github.com/XENONnT/straxen/pull/45
         # and https://github.com/AxFoundation/strax/pull/225
+        
+        """
         peaks = strax.split_peaks(
-            peaks, hitlets, radon_records, rlinks, self.to_pe,
+            peaks, hitlets, records, rlinks, self.to_pe,
             algorithm='natural_breaks',
             threshold=self.natural_breaks_threshold,
             split_low=True,
             filter_wing_width=self.config['peak_split_filter_wing_width'],
             min_area=self.config['peak_split_min_area'],
             do_iterations=self.config['peak_split_iterations'])
-
+        """
         # FIXME: Saturation correction using non-saturated channels
         
         # FIXME: Compute tight coincidence level.
         
-        if self.config['diagnose_sorting'] and len(radon_records):
-            assert np.diff(radon_records['time']).min(initial=1) >= 0, "Records not sorted"
+        if self.config['diagnose_sorting'] and len(records):
+            assert np.diff(records['time']).min(initial=1) >= 0, "Records not sorted"
             assert np.diff(hitlets['time']).min(initial=1) >= 0, "Hits/Hitlets not sorted"
             assert np.all(peaks['time'][1:]
                           >= strax.endtime(peaks)[:-1]), "Peaks not disjoint"    
@@ -199,8 +198,8 @@ class RadonPeaks(strax.Plugin):
         counts = strax.touching_windows(hitlets, peaks)
         counts = np.diff(counts, axis=1).flatten()
         peaks['n_hits'] = counts
-        return dict(radon_peaks=peaks,
-                    radon_lone_hits=lone_hits)
+        return dict(peaks=peaks,
+                    lone_hits=lone_hits)
 
     def natural_breaks_threshold(self, peaks):
         rise_time = -peaks['area_decile_from_midpoint'][:, 1]
@@ -429,18 +428,277 @@ def _peak_saturation_correction_inner(channel_saturated, records, p,
         help='Save (left, right) samples besides hits; cut the rest'),
         *HITFINDER_OPTIONS,
 )
-class RadonHits(strax.Plugin):
+class Hits(strax.Plugin):
     """
     Find hits using the find_hits algorithm in strax.
     """
-    __version__ = '0.0.7'
+    __version__ = '0.0.11'
 
     parallel = 'True'
     rechunk_on_save = False
-    depends_on = 'radon_records'
-    data_kind = 'radon_peaks'
+    depends_on = 'records'
+    data_kind = 'peaks'
     dtype = strax.hit_dtype
 
-    def compute(self, radon_records):
-        hits = strax.find_hits(radon_records, min_amplitude=amstrax.hit_min_amplitude(self.config['hit_min_amplitude']))
+    def compute(self, records):
+        print('in compute of hits test')
+        hits = strax.find_hits(records, min_amplitude=amstrax.hit_min_amplitude(self.config['hit_min_amplitude']))
         return hits
+    
+# # For n_competing, which is temporarily added to PeakBasics	
+# @export
+# @strax.takes_config(
+#     strax.Option('min_area_fraction', default=0.5,
+#                  help='The area of competing peaks must be at least '
+#                       'this fraction of that of the considered peak'),
+#     strax.Option('nearby_window', default=int(1e6),
+#                  help='Peaks starting within this time window (on either side)'
+#                       'in ns count as nearby.'),
+# )
+# class PeakBasics(strax.Plugin):
+#     provides = ('peak_basics',)
+#     depends_on = ('peaks')
+#     data_kind = ('peaks')
+
+#     parallel = 'False'
+#     rechunk_on_save = False
+#     __version__ = '0.1.7'
+#     dtype = [
+#         (('Start time of the peak (ns since unix epoch)',
+#           'time'), np.int64),
+#         (('End time of the peak (ns since unix epoch)',
+#           'endtime'), np.int64),
+#         (('Peak integral in PE',
+#           'area'), np.float32),
+#         (('Number of PMTs contributing to the peak',
+#           'n_channels'), np.int16),
+#         (('PMT number which contributes the most PE',
+#           'max_pmt'), np.int16),
+#         (('Area of signal in the largest-contributing PMT (PE)',
+#           'max_pmt_area'), np.int32),
+#         (('Width (in ns) of the central 50% area of the peak',
+#           'range_50p_area'), np.float32),
+#         (('Fraction of area seen by the top array',
+#           'area_fraction_top'), np.float32),
+#         (('Length of the peak waveform in samples',
+#           'length'), np.int32),
+#         (('Time resolution of the peak waveform in ns',
+#           'dt'), np.int16),
+#         ('n_competing', np.int32,  # temporarily due to chunking issues
+#          'Number of nearby larger or slightly smaller peaks')
+#     ]
+
+#     def compute(self, peaks):
+#         p = peaks
+#         p = strax.sort_by_time(p)
+#         r = np.zeros(len(p), self.dtype)
+#         for q in 'time length dt area'.split():
+#             r[q] = p[q]
+#         r['endtime'] = p['time'] + p['dt'] * p['length']
+#         r['n_channels'] = (p['area_per_channel'] > 0).sum(axis=1)
+#         r['range_50p_area'] = p['width'][:, 5]
+#         r['max_pmt'] = np.argmax(p['area_per_channel'], axis=1)
+#         r['max_pmt_area'] = np.max(p['area_per_channel'], axis=1)
+
+#         # area_top = p['area_per_channel'][:, :8].sum(axis=1)
+#         area_top = p['area_per_channel'][:, 1:2].sum(axis=1)  # top pmt in ch 1
+#         # Negative-area peaks get 0 AFT - TODO why not NaN?
+#         m = p['area'] > 0
+#         r['area_fraction_top'][m] = area_top[m] / p['area'][m]
+#         # n_competing temporarily due to chunking issues
+#         r['n_competing'] = self.find_n_competing(
+#             peaks,
+#             window=self.config['nearby_window'],
+#             fraction=self.config['min_area_fraction'])
+#         return r
+
+#     # n_competing
+#     def get_window_size(self):
+#         return 2 * self.config['nearby_window']
+
+#     @staticmethod
+#     @numba.jit(nopython=True, nogil=True, cache=False)
+#     def find_n_competing(peaks, window, fraction):
+#         n = len(peaks)
+#         t = peaks['time']
+#         a = peaks['area']
+#         results = np.zeros(n, dtype=np.int16)
+#         left_i = 0
+#         right_i = 0
+#         for i, peak in enumerate(peaks):
+#             while t[left_i] + window < t[i] and left_i < n - 1:
+#                 left_i += 1
+#             while t[right_i] - window < t[i] and right_i < n - 1:
+#                 right_i += 1
+#             results[i] = np.sum(a[left_i:right_i + 1] > a[i] * fraction)
+
+#         return results
+
+
+# @export
+# class PeakPositions(strax.Plugin):
+#     depends_on = ('peaks', 'peak_classification')
+#     rechunk_on_save = False
+#     __version__ = '0.0.34'  # .33 for LNLIKE
+#     dtype = [
+#         ('xr', np.float32,
+#          'Interaction x-position'),
+#         ('yr', np.float32,
+#          'Interaction y-position'),
+#         ('r', np.float32,
+#          'radial distance'),
+#         ('time', np.int64, 'Start time of the peak (ns since unix epoch)'),
+#         ('endtime', np.int64, 'End time of the peak (ns since unix epoch)')
+#     ]
+
+#     def setup(self):
+#         # z position of the in-plane SiPMs
+#         z_plane = 10
+#         # radius of the cylinder for SiPMs at the side
+#         r_cylinder = 22
+#         # radius of a SiPM - I assume circular SiPMs with a radius to make the area correspond to a 3x3mm2 square.
+#         r_sipm = 1.6925
+#         # build geometry
+#         geo = GeoParameters(z_plane=z_plane, r_cylinder=r_cylinder, r_sipm=r_sipm)
+
+#         sipm = SiPM(type="plane", position=[0, -15, z_plane], qeff=0.25)
+#         geo.add_sipm(sipm)
+#         sipm = SiPM(type="plane", position=[-13, -7.5, z_plane], qeff=0.25)
+#         geo.add_sipm(sipm)
+#         # sipm = SiPM(type="plane", position=[0, 15, z_plane], qeff=0.25)
+#         # geo.add_sipm(sipm)
+#         sipm = SiPM(type="plane", position=[13, -7.5, z_plane], qeff=0.25)
+#         geo.add_sipm(sipm)
+#         sipm = SiPM(type="plane", position=[-4, 0, z_plane], qeff=0.25)
+#         geo.add_sipm(sipm)
+#         sipm = SiPM(type="plane", position=[4, 0, z_plane], qeff=0.25)
+#         geo.add_sipm(sipm)
+#         sipm = SiPM(type="plane", position=[-13, 7.5, z_plane], qeff=0.25)
+#         geo.add_sipm(sipm)
+#         sipm = SiPM(type="plane", position=[13, 7.5, z_plane], qeff=0.25)
+#         geo.add_sipm(sipm)
+
+#         self.geo = geo
+
+#     def compute(self, peaks):
+
+#         result = np.empty(len(peaks), dtype=self.dtype)
+
+#         if not len(peaks):
+#             return result
+
+#         for ix, p in enumerate(peaks):
+
+#             if p['type'] != 2:
+#                 continue
+
+#             # if [X] channel is not working
+#             k = np.delete(p['area_per_channel'], [2])
+#             for i, area in enumerate(k):
+#                 self.geo.sipms[i].set_number_of_hits(area)
+
+#             # if all 8 channels are working
+#             # for i, area in enumerate(p['area_per_channel']):
+#             #     self.geo.sipms[i].set_number_of_hits(area)
+
+#             posrec = Reconstruction(self.geo)
+#             pos = posrec.reconstruct_position('CHI2')
+#             for key in ['xr', 'yr']:
+#                 result[key][ix] = pos[key]
+
+#             for q in ['time', 'endtime']:
+#                 result[q] = p[q]
+
+#         result['r'] = (result['xr'] ** 2 + result['yr'] ** 2) ** (1 / 2)
+#         return result
+
+
+# @export
+# @strax.takes_config(
+#     strax.Option('s1_max_width', default=60,
+#                  help="Maximum (IQR) width of S1s"),
+#     strax.Option('s1_min_area', default=4,
+#                  help="Minimum area (PE) for S1s"),
+#     strax.Option('s2_min_area', default=4,
+#                  help="Minimum area (PE) for S2s"),
+#     strax.Option('s2_min_width', default=100,
+#                  help="Minimum width for S2s"))
+# class PeakClassification(strax.Plugin):
+#     rechunk_on_save = False
+#     __version__ = '0.0.4'
+#     depends_on = ('peak_basics')
+#     dtype = [
+#         ('type', np.int8, 'Classification of the peak.'),
+#         ('time', np.int64, 'Start time of the peak (ns since unix epoch)'),
+#         ('endtime', np.int64, 'End time of the peak (ns since unix epoch)')
+#     ]
+#     parallel = True
+
+#     def compute(self, peaks):
+#         p = peaks
+#         r = np.zeros(len(p), dtype=self.dtype)
+
+#         is_s1 = p['area'] >= self.config['s1_min_area']
+#         is_s1 &= p['range_50p_area'] < self.config['s1_max_width']
+#         r['type'][is_s1] = 1
+
+#         is_s2 = p['area'] > self.config['s2_min_area']
+#         is_s2 &= p['range_50p_area'] > self.config['s2_min_width']
+#         r['type'][is_s2] = 2
+
+#         for q in ['time', 'endtime']:
+#             r[q] = p[q]
+
+#         return r
+
+# n_competing experiencing re-chunking issues, temporarily added to PeakBasics	        return results - 1
+# @export
+# @strax.takes_config(
+#     strax.Option('min_area_fraction', default=0.5,
+#                  help='The area of competing peaks must be at least '
+#                       'this fraction of that of the considered peak'),
+#     strax.Option('nearby_window', default=int(1e6),
+#                  help='Peaks starting within this time window (on either side)'
+#                       'in ns count as nearby.'),
+# )
+# class NCompeting(strax.OverlapWindowPlugin):   #from NCompetingTop
+#     depends_on = ('peak_basics',)           #from peak_basics_top
+#     rechunk_on_save = False
+#     dtype = [
+#         ('n_competing', np.int32,
+#             'Number of nearby larger or slightly smaller peaks'),
+#         ('time', np.int64, 'Start time of the peak (ns since unix epoch)'),
+#         ('endtime', np.int64, 'End time of the peak (ns since unix epoch)')
+#         ]
+#     __version__ = '0.0.27'
+#
+#     def get_window_size(self):
+#         return 2 * self.config['nearby_window']
+#
+#     def compute(self, peaks):
+#         results=np.zeros(len(peaks),dtype=self.dtype)
+#         results['time']=peaks['time']
+#         results['endtime']=peaks['endtime']
+#         results['n_competing']=self.find_n_competing(
+#             peaks,
+#             window=self.config['nearby_window'],
+#             fraction=self.config['min_area_fraction'])
+#         return results
+#
+#     @staticmethod
+#     @numba.jit(nopython=True, nogil=True, cache=False)
+#     def find_n_competing(peaks, window, fraction):
+#         n = len(peaks)
+#         t = peaks['time']
+#         a = peaks['area']
+#         results = np.zeros(n, dtype=np.int16)
+#         left_i = 0
+#         right_i = 0
+#         for i, peak in enumerate(peaks):
+#             while t[left_i] + window < t[i] and left_i < n - 1:
+#                 left_i += 1
+#             while t[right_i] - window < t[i] and right_i < n - 1:
+#                 right_i += 1
+#             results[i] = np.sum(a[left_i:right_i + 1] > a[i] * fraction)
+#
+#         return results
