+++ amstrax/plugins/raw_records/__init__.py
@@ -1,0 +1,2 @@
+from . import daqreader
+from .daqreader import *
