+++ amstrax/plugins/records/__init__.py
@@ -1,0 +1,9 @@
+from . import pulse_processing_alt_baseline
+from .pulse_processing_alt_baseline import *
+
+from . import pulse_processing
+from .pulse_processing import *
+
+from . import radon_pulse_processing
+from .radon_pulse_processing import *
+
